% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarExclude.R
\name{covarExclude}
\alias{covarExclude}
\title{Match Column Names to be Excluded}
\usage{
covarExclude(colname, exclude)
}
\arguments{
\item{colname}{column name.}

\item{exclude}{list of covariate name pairs or triplets to be excluded.}
}
\value{
Logical
}
\description{
Internal function called by \code{hbal} to serially expand covariates.
}
\author{
Yiqing Xu, Eddie Yang
}
