% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{.onAttach}
\alias{.onAttach}
\title{Subsidiary hbal Function}
\usage{
.onAttach(lib, pkg)
}
\arguments{
\item{lib}{libname}

\item{pkg}{package name}
}
\value{
No return value, called for the side effect of printing the package startup message.
}
\description{
Function to load package description.
}
\references{
Xu, Y., & Yang, E. (2022). Hierarchically Regularized Entropy Balancing. Political Analysis, 1-8. doi:10.1017/pan.2022.12
}
