% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fLM.R
\name{fLM}
\alias{fLM}
\title{Covariate Residualization}
\usage{
fLM(y, x)
}
\arguments{
\item{y}{x in RcppEigen::fastLm}

\item{x}{y in RcppEigen::fastLm}
}
\value{
Residualized covariates
}
\description{
Wrapper for RcppEigen's fastLm. Internal function called by \code{hbal} to residualize covariates.
}
\author{
Yiqing Xu, Eddie Yang
}
