% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResidual.R
\name{getResidual}
\alias{getResidual}
\title{Covariate Residualization}
\usage{
getResidual(P, pos.list = NULL)
}
\arguments{
\item{P}{covariate matrix}

\item{pos.list}{groupings of covariates}
}
\value{
Residualized covariates
}
\description{
Internal function called by \code{hbal} to residualize covariates.
}
\author{
Yiqing Xu, Eddie Yang
}
