% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbal-data.R
\docType{data}
\name{hbal-data}
\alias{hbal-data}
\title{Data from Black \& Owens (2016) and Hazlett (2020)}
\format{
The contenderJudges dataset has 2675 rows and 13 columns.
\describe{
  \item{presIdeoVote}{ideological alignment of judges' votes with the sitting President (outcome)}
  \item{treatFinal0}{treatment indicator for vacancy period}
  \item{judgeJCS}{judge’s Judicial Common Space (JCS)score}
  \item{presDist}{Ideological distribution of the sitting President}
  \item{panelDistJCS}{ideological composition of the panel with whom the judge sat}
  \item{circmed}{median JCS score of the circuit judges}
  \item{sctmed}{JCS score of the median justice on the Supreme Court}
  \item{coarevtc}{indicator for whether the case decision was reversed by the circuit court}
  \item{casepub}{indicator for the publication status of thecourt’s opinion}
  \item{judge}{name of the judge}
}

The LaLonde dataset has 10171 rows and 10 columns.
\describe{
  \item{nsw}{treatment indicator of whether an individual participated in the National Supported Work (NSW) program}
  \item{age}{}
  \item{educ}{years of education}
  \item{black}{demographic indicator variables for Black}
  \item{hisp}{idemographic indicator variables for Hispanic}
  \item{married}{demographic indicator variables for married}
  \item{re74}{real earnings in 1974}
  \item{re75}{real earnings in 1975}
  \item{re78}{real earnings in 1978, outcome}
  \item{u74}{unemployment indicator for 1974}
  \item{u75}{unemployment indicator for 1975}
  \item{u78}{unemployment indicator for 1978}
  \item{nodegr}{indicator for no high school degree}
}
}
\usage{
data(hbal)
}
\description{
The contenderJudges dataset is from Black \& Owens (2016): Courting the president: how circuit court judges alter their behavior for promotion to the Supreme Court
This dataset includes 10,171 period-judge observations for a total of 68 judges.
The treatment variable of interest is \code{treatFinal0}, which indicates whether there was a vacancy in the Supreme Court
The outcome of interest is ideological alignment of judges' votes with the sitting President (\code{presIdeoVote}).
The remaining variables are characteristics of the judges and courts, to be used as controls.

The LaLonde dataset has treated units from Dehejia and Wahba (1999), containing 185 individuals; data on the control units is from Panel Study
of Income Dynamics (PSID-1), containing 2,490 individuals.
}
\references{
\itemize{
  \item Black, R. C., & Owens, R. J. (2016). Courting the president: how circuit court judges alter their behavior for promotion to the Supreme Court. American Journal of Political Science, 60(1), 30-43.
  \item Dehejia, R. H., & Wahba, S. (1999). Causal effects in nonexperimental studies: Reevaluating the evaluation of training programs. Journal of the American statistical Association, 94(448), 1053-1062.
  \item Hazlett, C. (2020). KERNEL BALANCING. Statistica Sinica, 30(3), 1155-1189.
}
}
\keyword{datasets}
