% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{plot.hbal}
\alias{plot.hbal}
\alias{plot}
\title{Plotting Covariate Balance from an \code{hbal} Object}
\usage{
\method{plot}{hbal}(x, type = 'balance', log = TRUE, base_size = 10, ...)
}
\arguments{
\item{x}{an object of class \code{hbalobject} as returned by \code{hbal}.}

\item{type}{type of graph to plot.}

\item{log}{log scale for the weight plot}

\item{base_size}{base font size}

\item{...}{Further arguments to be passed to \code{plot.hbal()}.}
}
\value{
For \code{type = "weight"}, a \code{ggplot} object holding the histogram and
density of the control units' balancing weights (on the log scale when
\code{log = TRUE}); it is drawn when printed. For any other \code{type}, the
covariate-balance panels are drawn on the current graphics device and the
assembled \code{gtable} is returned invisibly.
}
\description{
This function plots the covariate difference between the control and treatment groups in standardized means before and after weighting.
}
\author{
Yiqing Xu, Eddie Yang
}
