% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{summary.hbal}
\alias{summary.hbal}
\alias{summary}
\title{Summarizing from an \code{hbal} Object}
\usage{
\method{summary}{hbal}(object, print.level = 0, ...)
}
\arguments{
\item{object}{an object of class \code{hbalobject} as returned by \code{hbal}.}

\item{print.level}{level of details to be printed}

\item{...}{Further arguments to be passed to \code{summary.hbal()}.}
}
\value{
Called for its side effect of printing a summary of the \code{hbal} object:
the call, the numbers of treated and control units, the covariate groups with
their penalties, and the balance table. The balance table
(\code{object$bal.tab}) is returned invisibly.
}
\description{
This function prints a summary from an \code{hbal} Object.
}
\author{
Yiqing Xu, Eddie Yang
}
