% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCoef.R
\name{updateCoef}
\alias{updateCoef}
\title{Update lambda}
\usage{
updateCoef(old.coef, new.coef, counter)
}
\arguments{
\item{old.coef}{previous coefficients}

\item{new.coef}{new coefficients}

\item{counter}{which fold in CV}
}
\value{
updated coefficients
}
\description{
Internal function called by \code{hbal} to residualize covariates.
}
\author{
Yiqing Xu, Eddie Yang
}
