% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{att}
\alias{att}
\title{Estimating the ATT from an hbal object}
\usage{
att(hbalobject, method="lm_robust", ...)
}
\arguments{
\item{hbalobject}{an object of class \code{hbal} as returned by \code{hbal}.}

\item{method}{estimation method for the ATT. Default is the Lin (2016) estimator.}

\item{...}{arguments passed to lm_lin or  lm_robust}
}
\value{
A matrix of estimates with their robust standard errors
}
\description{
\code{att} estimates the average treatment effect on the treated (ATT) from an 
hbal object returned by \code{hbal}.
}
\details{
This is a wrapper for \code{lm_robust} and \code{lm_lin} from the \link{estimatr} package.
}
\examples{
#EXAMPLE 1
#Given W, Y, X, calculate the ATT through hbal
set.seed(92092)
W <- sample(0:1, 1000, replace=TRUE)
X <- data.frame(X1=rnorm(1000), X2=rnorm(1000))
Y <- rnorm(1000)
out <- hbal(Treatment = W, Y = Y, X = X)
lin <- att(out, method="lin")
}
\author{
Yiqing Xu, Eddie Yang
}
