% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarExpand.R
\name{covarExpand}
\alias{covarExpand}
\title{Serial Expansion of Covariates}
\usage{
covarExpand(X, exp.degree = 3, treatment = NULL, exclude = NULL)
}
\arguments{
\item{X}{matrix of covariates.}

\item{exp.degree}{the degree of the polynomial.}

\item{treatment}{treatment indicator}

\item{exclude}{list of covariate name pairs or triplets to be excluded.}
}
\value{
A matrix of serially expanded covariates
}
\description{
Internal function called by \code{hbal} to serially expand covariates.
}
\author{
Yiqing Xu, Eddie Yang
}
