% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidate.R
\name{crossValidate}
\alias{crossValidate}
\alias{CrossValidate}
\title{Ridge Penalty Selection through Cross Validation}
\usage{
crossValidate(
  folds = NULL,
  alpha = NULL,
  treatment = NULL,
  fold.co = NULL,
  fold.tr = NULL,
  coefs = NULL,
  control = NULL,
  penalty = NULL,
  constraint.tolerance = NULL,
  print.level = NULL,
  base.weight = NULL,
  full.t = NULL,
  full.c = NULL,
  p = NULL,
  shuffle.treat = NULL
)
}
\arguments{
\item{folds}{number of folds to perform cross validation.}

\item{alpha}{alpha. Controls degree of regularization.}

\item{treatment}{covariate matrix for treatment group.}

\item{fold.co}{fold assignments for control units.}

\item{fold.tr}{fold assignments for treated units.}

\item{coefs}{starting coefficients (lambda).}

\item{control}{covariate matrix for control group.}

\item{penalty}{list of hierarchical penalties.}

\item{constraint.tolerance}{tolerance level for imbalance.}

\item{print.level}{details of printed output.}

\item{base.weight}{target weight distribution for the control units.}

\item{full.t}{(unresidualized) ovariate matrix for treatment group.}

\item{full.c}{(unresidualized) ovariate matrix for control group.}

\item{p}{positions to apply alpha}

\item{shuffle.treat}{whether to create folds for the treated units}
}
\value{
alpha, lambda
}
\description{
Internal function called by \code{hbal} to select ridge penalties through cross-validation.
}
\author{
Yiqing Xu, Eddie Yang
}
