% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleSelection.R
\name{doubleSelection}
\alias{doubleSelection}
\title{Double Selection}
\usage{
doubleSelection(X, W, Y, grouping)
}
\arguments{
\item{X}{covaraite matrix}

\item{W}{treatment indicator}

\item{Y}{outcome variable}

\item{grouping}{groupings of covariates}
}
\value{
resX, penalty.list, covar.keep
}
\description{
Internal function called by \code{hbal} to perform double selection.
}
\author{
Yiqing Xu, Eddie Yang
}
