% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{plot.hbal}
\alias{plot.hbal}
\alias{plot}
\title{Plotting Covariate Balance from an \code{hbal} Object}
\usage{
\method{plot}{hbal}(x, ...)
}
\arguments{
\item{x}{an object of class \code{hbalobject} as returned by \code{hbal}.}

\item{...}{Further arguments to be passed to \code{plot.hbal()}.}
}
\value{
A matrix of ggplots of covariate balance by group
}
\description{
This function plots the covariate difference between the control and treatment groups in standardized means before and after weighting.
}
\author{
Yiqing Xu, Eddie Yang
}
