% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{summary.hbal}
\alias{summary.hbal}
\alias{summary}
\title{Summarizing from an \code{hbal} Object}
\usage{
\method{summary}{hbal}(object, ...)
}
\arguments{
\item{object}{an object of class \code{hbalobject} as returned by \code{hbal}.}

\item{...}{Further arguments to be passed to \code{summary.hbal()}.}
}
\value{
a summary table
}
\description{
This function prints a summary from an \code{hbal} Object.
}
\author{
Yiqing Xu, Eddie Yang
}
